mcp/requirements.txt
psycopg2
faker
python-dotenv 